requirements.txt
aiogram
aiocron~=1.8
cachetools~=5.3.3
httpx~=0.27.0
openpyxl
pandas~=2.2.1
ping3~=4.0.7
psycopg[binary]~=3.2.10
psutil~=5.9.8
py-cpuinfo~=9.0.0
python-dotenv~=1.0.1
pytz~=2024.1
SQLAlchemy~=2.0.38
